chore: 폴더링 정리 (Repository, Domain 내부에 Core 디렉토리 생성)
