+++ FoodDiary/Data/Sources/Core/Extensions/Encodable+Extension.swift
