+++ FoodDiary/Data/Sources/Core/FoodClassifier/FoodClassificationResult.swift
