+++ FoodDiary/Data/Sources/Core/FoodPhoto/FoodPhotoFetcher.swift
