--- FoodDiary/Data/Sources/TokenManager/TokenManager.swift
@@ -1,11 +1,0 @@
-//
-//  TokenProvider.swift
-//  Data
-//
-//  Created by 강대훈 on 1/23/26.
-//
-
-import Foundation
-import Domain
-
-// TODO: TokenManager
