+++ FoodDiary/Domain/Sources/Core/Image/ImageAssetable.swift
