+++ FoodDiary/Domain/Sources/Core/Image/ImageCacheManageable.swift
